Moving built in evaluators to their own file
